requirements.txt
discord.py>=2.6.4
python-dotenv>=1.0.0
Pillow>=10.0.0
pet-pet-gif>=1.0.2
yt-dlp>=2024.1.1
aiohttp>=3.9.0
uwuify>=1.3.1
moviepy<2.0
deep-translator>=1.11.0
requests>=2.28.0
beautifulsoup4>=4.12.0
